created seperate rule for family level acmg_sf report
